refactor: breedList 파일 constants로 이동
